import requirements, refactor, hide buttons from user

+++ backend/api/projects/delete-project.php
@@ -1,13 +1,10 @@
 <?php
     require_once("../../db/db.php");
-    //require_once("../users/is-admin.php");
     session_start();
 
-    //if(isset($_SESSION["user"])){
+    if(isset($_SESSION["userId"]) && isset($_SESSION["userRoleId"]) && $_SESSION["userRoleId"] == 1) {
         $db = new DB();
-        //$adminCheck = new AdminCheck();
         $connection = $db->getConnection();
-        //if ($adminCheck->isAdmin($_SESSION["user"]["role_id"])) {
             $project_id = $_GET["projectId"];
             try {
                 $sql = "DELETE FROM projects WHERE id = :id";
@@ -20,11 +17,8 @@
                 http_response_code(500);
                 echo json_encode(["message" => $exc->getMessage()]);
             }
-    //     } else {
-    //         http_response_code(403);
-    //         echo json_encode(["message" => "Невалидни права за достъп - потребителят не е администратор"]);
-    //     }
-    // } else {
-    //     http_response_code(401);
-    // }
+        } else {
+            http_response_code(403);
+            echo json_encode(["message" => "Невалидни права за достъп - потребителят не е администратор"]);
+        }
 ?>
